refactor: add sabliercmd pkg
This will enable doc generation
